# src/main/openapi/oms_api-docs.yml
openapi: 3.0.3
info:
  title: Digital Assets OMS Command API
  version: 1.0.0
  description: Command API for submitting commands to oms

paths:
  /execute:
    post:
      summary: Execute an oms command
      operationId: executeCommand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Command'
            example:
              type: "OrderCreateCmd"
              order:
                id: "20250531-test-001"
                quantity: 100
      responses:
        '200':
          description: Command executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResult'
        '400':
          description: Invalid command

components:
  schemas:
    Command:
      type: object
      discriminator:
        propertyName: type
        mapping:
          QuoteRequestCreateCmd: '#/components/schemas/QuoteRequestCreateCmd'
          QuoteCreateCmd: '#/components/schemas/QuoteCreateCmd'
          OrderCreateCmd: '#/components/schemas/OrderCreateCmd'
          OrderAcceptCmd: '#/components/schemas/OrderAcceptCmd'
          ExecutionCreateCmd: '#/components/schemas/ExecutionCreateCmd'
          ExecutionWhackCmd: '#/components/schemas/ExecutionWhackCmd'
          ExecutionBustCmd: '#/components/schemas/ExecutionBustCmd'
      required:
        - type
      properties:
        type:
          type: string
          description: Command type discriminator
        version:
          type: string
          description: Command API version

    QuoteRequestCreateCmd:
      allOf:
        - $ref: '#/components/schemas/Command'
        - type: object
          properties:
            quoteRequest:
              $ref: '#/components/schemas/QuoteRequest'

    QuoteCreateCmd:
      allOf:
        - $ref: '#/components/schemas/Command'
        - type: object
          properties:
            quote:
              $ref: '#/components/schemas/Quote'

    OrderCreateCmd:
      allOf:
        - $ref: '#/components/schemas/Command'
        - type: object
          properties:
            order:
              $ref: '#/components/schemas/Order'

    OrderAcceptCmd:
      allOf:
        - $ref: '#/components/schemas/Command'
        - type: object
          properties:
            orderId:
              type: string

    ExecutionCreateCmd:
      allOf:
        - $ref: '#/components/schemas/Command'
        - type: object
          properties:
            execution:
              $ref: '#/components/schemas/Execution'

    ExecutionWhackCmd:
      allOf:
        - $ref: '#/components/schemas/Command'
        - type: object
          properties:
            executionId:
              type: string

    ExecutionBustCmd:
      allOf:
        - $ref: '#/components/schemas/Command'
        - type: object
          properties:
            executionId:
              type: string

    Quote:
      type: object
      properties:
        quoteRequestId:
          type: string
          description: ID of the associated quote request
        price:
          type: number
          description: Price of the quote

    QuoteRequest:
      type: object
      properties:
        symbol:
          type: string
        quantity:
          type: number
        currency:
          type: string

    Order:
      type: object
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        price:
          type: number
        quantity:
          type: number
        timeInForce:
          type: string
          enum: [GTC, IOC, FOK]

    Execution:
      type: object
      properties:
        execId:
          type: string
        orderId:
          type: string

    CommandResult:
      type: object
      properties:
        id:
          type: string
          description: unique identifier for object created by the command
        status:
          type: string
          enum: [OK, FAILED]
        message:
          type: string
          description: human-readable description of the command result

    CommandMessage:
      type: object
      properties:
        commandType:
          type: string
          description: The type of command contained in this message
          enum:
            - QuoteRequestCreateCmd
            - QuoteCreateCmd
            - OrderCreateCmd
            - OrderAcceptCmd
            - ExecutionCreateCmd
            - ExecutionWhackCmd
            - ExecutionBustCmd
        command:
          oneOf:
            - $ref: '#/components/schemas/QuoteRequestCreateCmd'
            - $ref: '#/components/schemas/QuoteCreateCmd'
            - $ref: '#/components/schemas/OrderCreateCmd'
            - $ref: '#/components/schemas/OrderAcceptCmd'
            - $ref: '#/components/schemas/ExecutionCreateCmd'
            - $ref: '#/components/schemas/ExecutionWhackCmd'
            - $ref: '#/components/schemas/ExecutionBustCmd'
      required:
        - commandType
        - command